Create Interface
Taken the code we found and put it together, added some minor changes to allow for a text log archive if we can get the rest of it to interface (see savelog()). Currently doesn't have the ability to send to the text box, researching how to code it to do so.